+++ Configuration/TCA/Overrides/tt_content.php
@@ -1,12 +1,17 @@
 <?php
 
-defined('TYPO3') or die();
+declare(strict_types=1);
 
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPlugin(
-    [
-        'LLL:EXT:oidc/Resources/Private/Language/locallang_db.xlf:tt_content.oidc_login',
-        'oidc_login',
-    ],
-    'list_type',
+use TYPO3\CMS\Core\Schema\Struct\SelectItem;
+use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
+
+ExtensionManagementUtility::addPlugin(
+    new SelectItem(
+        type: 'select',
+        label: 'LLL:EXT:oidc/Resources/Private/Language/locallang_db.xlf:tt_content.oidc_login',
+        value: 'oidc_login',
+        icon: 'ext-oidc-icon'
+    ),
+    'CType',
     'oidc'
 );