+++ ext_localconf.php
@@ -45,24 +45,6 @@ ExtensionManagementUtility::addService(
     ]
 );
 
-// Add typoscript for custom login plugin
-ExtensionManagementUtility::addTypoScript('oidc', 'setup', '
-# Setting oidc plugin TypoScript
-plugin.tx_oidc_login = USER_INT
-plugin.tx_oidc_login {
-    userFunc = Causal\Oidc\Controller\LoginController->login
-    defaultRedirectPid =
-    # Additional URL parameters for the authorization URL of the identity server
-    authorizationUrlOptions {
-        # login_theme = dark
-    }
-}
-');
-ExtensionManagementUtility::addTypoScript('oidc', 'setup', '
-# Setting oidc plugin TypoScript
-tt_content.list.20.oidc_login =< plugin.tx_oidc_login
-', 'defaultContentRendering');
-
 // Require 3rd-party libraries, in case TYPO3 does not run in composer mode
 $pharFileName = ExtensionManagementUtility::extPath('oidc') . 'Libraries/league-oauth2-client.phar';
 if (is_file($pharFileName)) {